Moved Currency interface + entity back to Localization folder
