+++ src/Core/Localization/Currency/Repository.php
@@ -26,6 +26,7 @@
 
 namespace PrestaShop\PrestaShop\Core\Localization\Currency;
 
+use PrestaShop\PrestaShop\Core\Localization\Currency;
 use PrestaShop\PrestaShop\Core\Localization\Currency\RepositoryInterface as CurrencyRepositoryInterface;
 use PrestaShop\PrestaShop\Core\Localization\Currency\DataRepositoryInterface as CurrencyDataRepositoryInterface;
